return the deleted/patched resource

+++ src/1-data-providers/storage/storage-todos.ts
@@ -68,34 +68,42 @@ export const storageTodos = ({ storageClient, logger }: { storageClient: Storage
 		return todo;
 	};
 
-	const patchById = async (id: Todo["id"], partialTodo: Partial<TodoWithoutId>): Promise<boolean> => {
+	const patchById = async (id: Todo["id"], partialTodo: Partial<TodoWithoutId>): Promise<Todo | null> => {
 		logger.verbose(`updating todo ${id}`);
+
 		await storageClient.mutation(
 			`
 				UPDATE todos
 				SET text = COALESCE($text, text),
 					completed = COALESCE($completed, completed)
-				WHERE id = $id
+				WHERE id = $id;
 			`,
 			{ $id: id, $text: partialTodo.text ?? null, $completed: partialTodo.completed ?? null }
 		);
 
-		const data = await storageClient.queryOne(
-			z.object({ rowsUpdated: z.number() }).strict(),
-			`SELECT changes() AS rowsUpdated`
-		);
-		const updated = data ? data.rowsUpdated > 0 : false;
-		logger.verbose(updated ? `updating todo ${id} done` : `updating todo ${id} failed, todo not found`);
+		const todo = await getById(id);
+		if (!todo) {
+			logger.verbose(`updating todo ${id} failed, todo not found`);
+			return null;
+		}
 
-		return updated;
+		logger.verbose(`updating todo ${id} done`);
+		return todo;
 	};
 
 	const deleteAll = async (): Promise<void> => {
 		logger.verbose(`deleting all todos`);
 		await storageClient.mutation(`DELETE FROM todos`);
 	};
-	const deleteById = async (id: Todo["id"]): Promise<boolean> => {
+	const deleteById = async (id: Todo["id"]): Promise<Todo | null> => {
 		logger.verbose(`deleting todo ${id}`);
+
+		const todo = await getById(id);
+		if (!todo) {
+			logger.verbose(`deleting todo ${id} failed, todo not found`);
+			return null;
+		}
+
 		await storageClient.mutation(
 			`
 				DELETE FROM todos
@@ -103,15 +111,9 @@ export const storageTodos = ({ storageClient, logger }: { storageClient: Storage
 			`,
 			{ $id: id }
 		);
+		logger.verbose(`deleting todo ${id} done`);
 
-		const data = await storageClient.queryOne(
-			z.object({ rowsDeleted: z.number() }).strict(),
-			`SELECT changes() AS rowsDeleted`
-		);
-		const deleted = data ? data.rowsDeleted > 0 : false;
-		logger.verbose(deleted ? `deleting todo ${id} done` : `deleting todo ${id} failed, todo not found`);
-
-		return deleted;
+		return todo;
 	};
 
 	return {
